requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
Pillow~=8.1.2
pyTelegramBotAPI==4.2.1
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six~=1.15.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
cffi~=1.14.5
future~=0.18.2
olefile~=0.46
pip~=21.3.1
toml~=0.10.2
wheel~=0.34.2
cryptography~=3.4.6
lxml~=4.6.3
keyring~=18.0.1
colorama~=0.4.3
chardet~=3.0.4
simplejson~=3.16.0
update~=0.0.1
setuptools~=54.0.0
SQLAlchemy~=1.4.2
python-dateutil~=2.7.3